+++ Source/MulticopterSim/SimFlightController.h
@@ -24,7 +24,7 @@ public:
 
     void init(uint8_t  axismap[5], uint8_t buttonmap[3], bool reversedVerticals, bool springyThrottle, bool useButtonForAux);
 
-    void update(float axes[6], uint8_t buttons, float quat[4], float gyro[3]);
+    void update(int32_t axes[6], uint8_t buttons, float quat[4], float gyro[3]);
 
     static SimFlightController * createSimFlightController(void);
 };
